Moved class to server directory
